requirements.txt
httpx~=0.27
python-dotenv~=1.0
# numpy~=1.22.4
pyserial~=3.5
RPi.GPIO~=0.7.1
